// src/types/media.type.ts
export type TMedia = {
  width: number;
  height: number;
  mime: string;
  size: number;
  path: string;
  modificationDate: string;
};

export type TResponseUploadImage = {
  asset_id: string;
  public_id: string;
  version: number;
  version_id: string;
  signature: string;
  width: number;
  height: number;
  format: string;
  resource_type: string;
  created_at: Date;
  tags: any[];
  bytes: number;
  type: string;
  etag: string;
  placeholder: boolean;
  url: string;
  secure_url: string;
  folder: string;
  access_mode: string;
  original_filename: string;
};